+++ sdk/security/keyvault/azadmin/build.go
@@ -4,9 +4,6 @@
 //go:generate go run ./internal/generate/transforms.go
 //go:generate gofmt -w .
 //go:generate rm ./backup/constants.go
-//go:generate rm ./backup/go.mod
-//go:generate rm ./rbac/go.mod
-//go:generate rm ./settings/go.mod
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.